+++ frontend/techpick/src/providers/ThemeProvider.tsx
@@ -1,9 +1,6 @@
 'use client';
 import { useThemeStore } from '@/lib/stores/themeStore';
-import { darkTheme } from '@/shared/themes/darkTheme';
-import { lightTheme } from '@/shared/themes/lightTheme';
-import { commonTheme } from '@/shared/themes/commonTheme';
-
+import { lightTheme, darkTheme, commonTheme } from '@/shared/themes';
 export const ThemeProvider = ({ children }: { children: React.ReactNode }) => {
   const { isDarkMode } = useThemeStore();
