+++ frontend/techpick/src/shared/themes/commonThemeContract.css.ts
@@ -1,0 +1,34 @@
+import { createThemeContract } from '@vanilla-extract/css';
+
+export const commonThemeContract = createThemeContract({
+  space: {
+    xs: '',
+    sm: '',
+    md: '',
+    lg: '',
+    xl: '',
+  },
+  typography: {
+    fontFamily: '',
+    fontColor: '',
+    fontSize: {
+      xs: '',
+      sm: '',
+      md: '',
+      lg: '',
+      xl: '',
+    },
+    lineHeight: {
+      xs: '',
+      sm: '',
+      md: '',
+      lg: '',
+      xl: '',
+    },
+  },
+  breakpoints: {
+    mobile: '',
+    tablet: '',
+    desktop: '',
+  },
+});
