+++ frontend/techpick/src/shared/themes/index.ts
@@ -1,4 +1,5 @@
-export { commonTheme } from './commonTheme';
+export { commonTheme } from './commonTheme.css';
+export { commonThemeContract } from './commonThemeContract.css';
 export { themeContract } from './themeContract.css';
-export { darkTheme } from './darkTheme';
-export { lightTheme } from './lightTheme';
+export { darkTheme } from './darkTheme.css';
+export { lightTheme } from './lightTheme.css';
