+++ frontend/techpick/src/shared/themes/lightTheme.css.ts
