+++ frontend/techpick/src/shared/themes/themeContract.css.ts
@@ -9,36 +9,3 @@ export const themeContract = createThemeContract({
     font: '',
   },
 });
-
-export const commonThemeContract = createThemeContract({
-  space: {
-    xs: '',
-    sm: '',
-    md: '',
-    lg: '',
-    xl: '',
-  },
-  typography: {
-    fontFamily: '',
-    fontColor: '',
-    fontSize: {
-      xs: '',
-      sm: '',
-      md: '',
-      lg: '',
-      xl: '',
-    },
-    lineHeight: {
-      xs: '',
-      sm: '',
-      md: '',
-      lg: '',
-      xl: '',
-    },
-  },
-  breakpoints: {
-    mobile: '',
-    tablet: '',
-    desktop: '',
-  },
-});
